doc: stylish distributed-data.md, #24717
* move to docs package
* style changes
